+++ portal-impl/src/com/liferay/portal/verify/VerifyAuditedModel.java
@@ -16,6 +16,9 @@ package com.liferay.portal.verify;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.bean.PortalBeanLocatorUtil;
+import com.liferay.portal.kernel.dao.db.DBInspector;
+import com.liferay.portal.kernel.dao.db.DBType;
+import com.liferay.portal.kernel.dao.db.DBTypeToSQLMap;
 import com.liferay.portal.kernel.dao.jdbc.AutoBatchPreparedStatementUtil;
 import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.log.Log;
@@ -23,19 +26,23 @@ import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.security.auth.FullNameGenerator;
 import com.liferay.portal.kernel.security.auth.FullNameGeneratorFactory;
 import com.liferay.portal.kernel.util.LoggingTimer;
+import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.verify.model.VerifiableAuditedModel;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.sql.Timestamp;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
+import java.util.Set;
 import java.util.concurrent.Callable;
 
 /**
@@ -47,7 +54,7 @@ public class VerifyAuditedModel extends VerifyProcess {
 	public void verify(VerifiableAuditedModel... verifiableAuditedModels)
 		throws Exception {
 
-		List<String> unverifiedTableNames = new ArrayList<>();
+		Set<String> unverifiedTableNames = new HashSet<>();
 
 		for (VerifiableAuditedModel verifiableAuditedModel :
 				verifiableAuditedModels) {
@@ -55,19 +62,24 @@ public class VerifyAuditedModel extends VerifyProcess {
 			unverifiedTableNames.add(verifiableAuditedModel.getTableName());
 		}
 
-		List<VerifyAuditedModelCallable> verifyAuditedModelCallables =
-			new ArrayList<>(unverifiedTableNames.size());
-
 		while (!unverifiedTableNames.isEmpty()) {
-			int count = unverifiedTableNames.size();
+			List<VerifyAuditedModelCallable> verifyAuditedModelCallables =
+				new ArrayList<>(unverifiedTableNames.size());
 
 			for (VerifiableAuditedModel verifiableAuditedModel :
 					verifiableAuditedModels) {
 
-				if (unverifiedTableNames.contains(
-						verifiableAuditedModel.getJoinByTableName()) ||
-					!unverifiedTableNames.contains(
-						verifiableAuditedModel.getTableName())) {
+				String tableName = verifiableAuditedModel.getTableName();
+
+				if (!unverifiedTableNames.contains(tableName)) {
+					continue;
+				}
+
+				String relatedModelName =
+					verifiableAuditedModel.getRelatedModelName();
+
+				if ((relatedModelName != null) &&
+					unverifiedTableNames.contains(relatedModelName)) {
 
 					continue;
 				}
@@ -77,17 +89,16 @@ public class VerifyAuditedModel extends VerifyProcess {
 
 				verifyAuditedModelCallables.add(verifyAuditedModelCallable);
 
-				unverifiedTableNames.remove(
-					verifiableAuditedModel.getTableName());
+				unverifiedTableNames.remove(tableName);
 			}
 
-			if (unverifiedTableNames.size() == count) {
+			if (verifyAuditedModelCallables.isEmpty()) {
 				throw new VerifyException(
 					"Circular dependency detected " + unverifiedTableNames);
 			}
-		}
 
-		doVerify(verifyAuditedModelCallables);
+			doVerify(verifyAuditedModelCallables);
+		}
 	}
 
 	@Override
@@ -101,55 +112,6 @@ public class VerifyAuditedModel extends VerifyProcess {
 		verify(
 			verifiableAuditedModels.toArray(
 				new VerifiableAuditedModel[verifiableAuditedModels.size()]));
-	}
-
-	protected Object[] getAuditedModelArray(
-			Connection con, String tableName, String pkColumnName, long primKey,
-			boolean allowAnonymousUser, long previousUserId)
-		throws Exception {
-
-		try (PreparedStatement ps = con.prepareStatement(
-				StringBundler.concat(
-					"select companyId, userId, createDate, modifiedDate from ",
-					tableName, " where ", pkColumnName, " = ?"))) {
-
-			ps.setLong(1, primKey);
-
-			try (ResultSet rs = ps.executeQuery()) {
-				if (rs.next()) {
-					long companyId = rs.getLong("companyId");
-
-					long userId = 0;
-					String userName = null;
-
-					if (allowAnonymousUser) {
-						userId = previousUserId;
-						userName = "Anonymous";
-					}
-					else {
-						userId = rs.getLong("userId");
-
-						userName = getUserName(con, userId);
-					}
-
-					Timestamp createDate = rs.getTimestamp("createDate");
-					Timestamp modifiedDate = rs.getTimestamp("modifiedDate");
-
-					return new Object[] {
-						companyId, userId, userName, createDate, modifiedDate
-					};
-				}
-
-				if (_log.isDebugEnabled()) {
-					_log.debug(
-						StringBundler.concat(
-							"Unable to find ", tableName, " ",
-							String.valueOf(primKey)));
-				}
-
-				return null;
-			}
-		}
 	}
 
 	protected Object[] getDefaultUserArray(Connection con, long companyId)
@@ -265,96 +227,192 @@ public class VerifyAuditedModel extends VerifyProcess {
 		throws Exception {
 
 		try (LoggingTimer loggingTimer = new LoggingTimer(
-				verifiableAuditedModel.getTableName())) {
+				verifiableAuditedModel.getTableName());
+			Connection con = DataAccess.getConnection()) {
 
-			StringBundler sb = new StringBundler(8);
+			String relatedModelName =
+				verifiableAuditedModel.getRelatedModelName();
 
-			sb.append("select ");
-			sb.append(verifiableAuditedModel.getPrimaryKeyColumnName());
-			sb.append(", companyId, userId");
+			if (relatedModelName != null) {
+				_verifyFromRelatedModel(con, verifiableAuditedModel, "userId");
 
-			if (verifiableAuditedModel.getJoinByTableName() != null) {
-				sb.append(StringPool.COMMA_AND_SPACE);
-				sb.append(verifiableAuditedModel.getJoinByTableName());
-			}
+				if (verifiableAuditedModel.isUpdateDates()) {
+					_verifyFromRelatedModel(
+						con, verifiableAuditedModel, "createDate");
 
-			sb.append(" from ");
-			sb.append(verifiableAuditedModel.getTableName());
-			sb.append(" where userName is null order by companyId");
-
-			Object[] auditedModelArray = null;
-
-			long previousCompanyId = 0;
-
-			try (Connection con = DataAccess.getConnection();
-				PreparedStatement ps1 = con.prepareStatement(sb.toString());
-				ResultSet rs = ps1.executeQuery();
-				PreparedStatement ps2 =
-					AutoBatchPreparedStatementUtil.autoBatch(
-						_createPreparedStatement(
-							con, verifiableAuditedModel.getTableName(),
-							verifiableAuditedModel.getPrimaryKeyColumnName(),
-							verifiableAuditedModel.isUpdateDates()))) {
-
-				while (rs.next()) {
-					long companyId = rs.getLong("companyId");
-
-					if (verifiableAuditedModel.getJoinByTableName() != null) {
-						long previousUserId = rs.getLong("userId");
-						long relatedPrimKey = rs.getLong(
-							verifiableAuditedModel.getJoinByTableName());
-
-						auditedModelArray = getAuditedModelArray(
-							con, verifiableAuditedModel.getRelatedModelName(),
-							verifiableAuditedModel.getRelatedPKColumnName(),
-							relatedPrimKey,
-							verifiableAuditedModel.isAnonymousUserAllowed(),
-							previousUserId);
-					}
-					else if (previousCompanyId != companyId) {
-						auditedModelArray = getDefaultUserArray(con, companyId);
-
-						previousCompanyId = companyId;
-					}
-
-					if (auditedModelArray == null) {
-						continue;
-					}
-
-					long primKey = rs.getLong(
-						verifiableAuditedModel.getPrimaryKeyColumnName());
-
-					verifyAuditedModel(
-						con, ps2, verifiableAuditedModel.getTableName(),
-						primKey, auditedModelArray,
-						verifiableAuditedModel.isUpdateDates());
+					_verifyFromRelatedModel(
+						con, verifiableAuditedModel, "modifiedDate");
 				}
 
-				ps2.executeBatch();
+				DBInspector dbInspector = new DBInspector(con);
+
+				if (dbInspector.hasColumn(relatedModelName, "userName")) {
+					_verifyFromRelatedModel(
+						con, verifiableAuditedModel, "userName");
+				}
 			}
+
+			_verifyUserIdRegularUsers(con, verifiableAuditedModel);
+
+			_verifyUnresolvedUsers(con, verifiableAuditedModel);
 		}
 	}
 
-	private PreparedStatement _createPreparedStatement(
-			Connection con, String tableName, String primaryKeyColumnName,
-			boolean updateDates)
-		throws SQLException {
+	private void _verifyFromRelatedModel(
+			Connection con, VerifiableAuditedModel verifiableAuditedModel,
+			String columnName)
+		throws Exception {
 
-		StringBundler sb = new StringBundler(7);
+		String sql = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(), " set ",
+			columnName, " = (select ", columnName, " from ",
+			verifiableAuditedModel.getRelatedModelName(), " where ",
+			verifiableAuditedModel.getTableName(), StringPool.PERIOD,
+			verifiableAuditedModel.getJoinByTableName(), " = ",
+			verifiableAuditedModel.getRelatedModelName(), StringPool.PERIOD,
+			verifiableAuditedModel.getRelatedPKColumnName(),
+			") where userName is null");
 
-		sb.append("update ");
-		sb.append(tableName);
-		sb.append(" set companyId = ?, userId = ?, userName = ?");
+		DBTypeToSQLMap dbTypeToSQLMap = new DBTypeToSQLMap(sql);
 
-		if (updateDates) {
-			sb.append(", createDate = ?, modifiedDate = ?");
+		sql = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(), " inner join ",
+			verifiableAuditedModel.getRelatedModelName(), " on ",
+			verifiableAuditedModel.getTableName(), StringPool.PERIOD,
+			verifiableAuditedModel.getJoinByTableName(), " = ",
+			verifiableAuditedModel.getRelatedModelName(), StringPool.PERIOD,
+			verifiableAuditedModel.getRelatedPKColumnName(), " set ",
+			verifiableAuditedModel.getTableName(), StringPool.PERIOD,
+			columnName, " = ", verifiableAuditedModel.getRelatedModelName(),
+			StringPool.PERIOD, columnName, " where ",
+			verifiableAuditedModel.getTableName(), ".userName is null");
+
+		dbTypeToSQLMap.add(DBType.MARIADB, sql);
+		dbTypeToSQLMap.add(DBType.MYSQL, sql);
+
+		runSQL(con, dbTypeToSQLMap);
+	}
+
+	private void _verifyUnresolvedUsers(
+			Connection con, VerifiableAuditedModel verifiableAuditedModel)
+		throws Exception {
+
+		String createDateSQL = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(),
+			" set createDate = ? where createDate is null");
+
+		String modifiedDateSQL = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(),
+			" set modifiedDate = ? where modifiedDate is null");
+
+		String modifyUserId = StringPool.BLANK;
+
+		if (!verifiableAuditedModel.isAnonymousUserAllowed()) {
+			modifyUserId = " userId = ?,";
 		}
 
-		sb.append(" where ");
-		sb.append(primaryKeyColumnName);
-		sb.append(" = ?");
+		String userNameSQL = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(), " set",
+			modifyUserId, " userName = ? where userName is null");
 
-		return con.prepareStatement(sb.toString());
+		try (PreparedStatement ps1 = AutoBatchPreparedStatementUtil.autoBatch(
+				con.prepareStatement(createDateSQL));
+			PreparedStatement ps2 = AutoBatchPreparedStatementUtil.autoBatch(
+				con.prepareStatement(modifiedDateSQL));
+			PreparedStatement ps3 = AutoBatchPreparedStatementUtil.autoBatch(
+				con.prepareStatement(userNameSQL))) {
+
+			if (verifiableAuditedModel.isAnonymousUserAllowed()) {
+				if (verifiableAuditedModel.isUpdateDates()) {
+					Timestamp auditDate = new Timestamp(
+						System.currentTimeMillis());
+
+					ps1.setTimestamp(1, auditDate);
+
+					ps1.addBatch();
+
+					ps2.setTimestamp(1, auditDate);
+
+					ps2.addBatch();
+				}
+
+				ps3.setString(1, "Anonymous");
+
+				ps3.addBatch();
+			}
+			else {
+				for (long companyId : PortalUtil.getCompanyIds()) {
+					Object[] defaultUserArray = getDefaultUserArray(
+						con, companyId);
+
+					if (verifiableAuditedModel.isUpdateDates()) {
+						Timestamp createDate = (Timestamp)defaultUserArray[3];
+
+						ps1.setTimestamp(1, createDate);
+
+						ps1.addBatch();
+
+						Timestamp modifiedDate = (Timestamp)defaultUserArray[4];
+
+						ps2.setTimestamp(1, modifiedDate);
+
+						ps2.addBatch();
+					}
+
+					long userId = (Long)defaultUserArray[1];
+					String defaultUserFullName = (String)defaultUserArray[2];
+
+					ps3.setLong(1, userId);
+					ps3.setString(2, defaultUserFullName);
+
+					ps3.addBatch();
+				}
+			}
+
+			ps1.executeBatch();
+			ps2.executeBatch();
+			ps3.executeBatch();
+		}
+	}
+
+	private void _verifyUserIdRegularUsers(
+			Connection con, VerifiableAuditedModel verifiableAuditedModel)
+		throws Exception {
+
+		String selectQuery = StringBundler.concat(
+			"select distinct userId from ",
+			verifiableAuditedModel.getTableName(), " where userName is null");
+
+		String updateQuery = StringBundler.concat(
+			"update ", verifiableAuditedModel.getTableName(),
+			" set userName = ? where userId = ?");
+
+		try (PreparedStatement ps1 = con.prepareStatement(selectQuery);
+			ResultSet rs = ps1.executeQuery();
+			PreparedStatement ps2 = AutoBatchPreparedStatementUtil.autoBatch(
+				con.prepareStatement(updateQuery))) {
+
+			Map<Long, String> fullNames = new HashMap<>();
+
+			while (rs.next()) {
+				long userId = rs.getLong("userId");
+
+				String fullName = getUserName(con, userId);
+
+				if (Validator.isNotNull(fullName)) {
+					fullNames.put(userId, fullName);
+				}
+			}
+
+			for (Map.Entry<Long, String> entry : fullNames.entrySet()) {
+				ps2.setString(1, entry.getValue());
+				ps2.setLong(2, entry.getKey());
+
+				ps2.addBatch();
+			}
+
+			ps2.executeBatch();
+		}
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(